# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beneficiaries API",
    "description": "Backend contract for the Arabic Beneficiaries mobile application. Authenticated endpoints use an HTTP Bearer access token. Business-controlled values such as points, attendance QR values, submission status, and duplicate awards are calculated by the server.",
    "contact": { "name": "Beneficiaries application team" },
    "license": { "name": "Proprietary" },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/questions/categories": {
      "get": {
        "tags": ["Questions"],
        "summary": "List active question-bank categories",
        "operationId": "list_categories_api_v1_questions_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuestionCategoryResponse"
                  },
                  "type": "array",
                  "title": "Response List Categories Api V1 Questions Categories Get"
                }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/v1/questions": {
      "get": {
        "tags": ["Questions"],
        "summary": "List published questions in one category",
        "operationId": "list_questions_api_v1_questions_get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Categoryid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionAnswerResponse"
                  },
                  "title": "Response List Questions Api V1 Questions Get"
                }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/questions/proposals": {
      "post": {
        "tags": ["Questions"],
        "summary": "Propose a question for servant review",
        "operationId": "propose_question_api_v1_questions_proposals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionProposalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/v1/questions/know-me": {
      "get": {
        "tags": ["Questions"],
        "summary": "List recent Know Me questions and answer status",
        "operationId": "list_know_me_questions_api_v1_questions_know_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowMeQuestionResponse"
                  },
                  "type": "array",
                  "title": "Response List Know Me Questions Api V1 Questions Know Me Get"
                }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/v1/questions/know-me/{question_id}/answers": {
      "post": {
        "tags": ["Questions"],
        "summary": "Answer a Know Me question and award eligible points once",
        "operationId": "submit_know_me_answer_api_v1_questions_know_me__question_id__answers_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Question Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/KnowMeAnswerRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubmissionResult" }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "message": { "type": "string", "title": "Message" },
          "code": { "type": "string", "title": "Code" },
          "fieldErrors": {
            "anyOf": [
              {
                "additionalProperties": { "type": "string" },
                "type": "object"
              },
              { "type": "null" }
            ],
            "title": "Fielderrors"
          }
        },
        "type": "object",
        "required": ["message", "code"],
        "title": "ApiError"
      },
...
      "KnowMeAnswerRequest": {
        "properties": {
          "answer": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Answer"
          }
        },
        "type": "object",
        "required": ["answer"],
        "title": "KnowMeAnswerRequest"
      },
      "KnowMeQuestionResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "label": { "type": "string", "title": "Label" },
          "askedAt": { "type": "string", "title": "Askedat" },
          "isToday": { "type": "boolean", "title": "Istoday" },
          "points": { "type": "integer", "title": "Points" },
          "answered": {
            "type": "boolean",
            "title": "Answered",
            "default": false
          }
        },
        "type": "object",
        "required": ["id", "label", "askedAt", "isToday", "points"],
        "title": "KnowMeQuestionResponse"
      },
...
      "ProposalStatus": {
        "type": "string",
        "enum": ["pending", "approved", "rejected"],
        "title": "ProposalStatus"
      },
      "QuestionAnswerResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "categoryId": { "type": "string", "title": "Categoryid" },
          "question": { "type": "string", "title": "Question" },
          "answer": { "type": "string", "title": "Answer" }
        },
        "type": "object",
        "required": ["id", "categoryId", "question", "answer"],
        "title": "QuestionAnswerResponse"
      },
      "QuestionCategoryResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "title": { "type": "string", "title": "Title" },
          "icon": { "type": "string", "title": "Icon" },
          "color": { "type": "string", "title": "Color" }
        },
        "type": "object",
        "required": ["id", "title", "icon", "color"],
        "title": "QuestionCategoryResponse"
      },
      "QuestionProposalRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Question"
          }
        },
        "type": "object",
        "required": ["question"],
        "title": "QuestionProposalRequest"
      },
      "QuestionProposalResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "status": { "$ref": "#/components/schemas/ProposalStatus" },
          "message": { "type": "string", "title": "Message" }
        },
        "type": "object",
        "required": ["id", "status", "message"],
        "title": "QuestionProposalResponse"
      },
...
      "SubmissionResult": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "pointsAwarded": { "type": "integer", "title": "Pointsawarded" },
          "totalPoints": { "type": "integer", "title": "Totalpoints" },
          "message": { "type": "string", "title": "Message" }
        },
        "type": "object",
        "required": ["id", "pointsAwarded", "totalPoints", "message"],
        "title": "SubmissionResult"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}